Simplifying helper methods.
